refactor LunchService
